Rename general utils to shared for clarity and fix paths
