+++ src/shared/global_constants.py
